Implement BPP

+++ src/main/java/com/epam/training/facade/BookingFacade.java
@@ -113,7 +113,7 @@ public interface BookingFacade {
      * @return Booked ticket object.
      * @throws java.lang.IllegalStateException if this place has already been booked.
      */
-    Ticket bookTicket(Integer userId, long eventId, int place, Category category);
+    Ticket bookTicket(Integer userId, Integer eventId, int place, Category category);
 
     /**
      * Get all booked tickets for specified user. Tickets should be sorted by event date in descending order.
